Refactor BoxHeader
