+++ src/atoms/mdhd.rs
@@ -3,6 +3,7 @@ use std::char::{decode_utf16, REPLACEMENT_CHARACTER};
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 
 use crate::*;
+use crate::atoms::*;
 
 
 #[derive(Debug, PartialEq)]
@@ -92,7 +93,7 @@ impl<R: Read + Seek> ReadBox<&mut BufReader<R>> for MdhdBox {
 impl<W: Write> WriteBox<&mut BufWriter<W>> for MdhdBox {
     fn write_box(&self, writer: &mut BufWriter<W>) -> Result<u64> {
         let size = self.box_size();
-        BoxHeader::new(Self::box_type(), size).write_box(writer)?;
+        BoxHeader::new(Self::box_type(), size).write(writer)?;
 
         write_box_header_ext(writer, self.version, self.flags)?;
 
@@ -143,7 +144,7 @@ fn get_language_code(language: &str) -> u16 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::read_box_header;
+    use crate::atoms::BoxHeader;
     use std::io::Cursor;
 
     fn test_language_code(lang: &str) {
@@ -179,7 +180,7 @@ mod tests {
 
         {
             let mut reader = BufReader::new(Cursor::new(&buf));
-            let header = read_box_header(&mut reader).unwrap();
+            let header = BoxHeader::read(&mut reader).unwrap();
             assert_eq!(header.name, BoxType::MdhdBox);
             assert_eq!(src_box.box_size(), header.size);
 
@@ -209,7 +210,7 @@ mod tests {
 
         {
             let mut reader = BufReader::new(Cursor::new(&buf));
-            let header = read_box_header(&mut reader).unwrap();
+            let header = BoxHeader::read(&mut reader).unwrap();
             assert_eq!(header.name, BoxType::MdhdBox);
             assert_eq!(src_box.box_size(), header.size);
