+++ src/atoms/stts.rs
@@ -2,6 +2,7 @@ use std::io::{BufReader, Seek, Read, BufWriter, Write};
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 
 use crate::*;
+use crate::atoms::*;
 
 
 #[derive(Debug, Default, PartialEq)]
@@ -56,7 +57,7 @@ impl<R: Read + Seek> ReadBox<&mut BufReader<R>> for SttsBox {
 impl<W: Write> WriteBox<&mut BufWriter<W>> for SttsBox {
     fn write_box(&self, writer: &mut BufWriter<W>) -> Result<u64> {
         let size = self.box_size();
-        BoxHeader::new(Self::box_type(), size).write_box(writer)?;
+        BoxHeader::new(Self::box_type(), size).write(writer)?;
 
         write_box_header_ext(writer, self.version, self.flags)?;
 
@@ -73,7 +74,7 @@ impl<W: Write> WriteBox<&mut BufWriter<W>> for SttsBox {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::read_box_header;
+    use crate::atoms::BoxHeader;
     use std::io::Cursor;
 
     #[test]
@@ -95,7 +96,7 @@ mod tests {
 
         {
             let mut reader = BufReader::new(Cursor::new(&buf));
-            let header = read_box_header(&mut reader).unwrap();
+            let header = BoxHeader::read(&mut reader).unwrap();
             assert_eq!(header.name, BoxType::SttsBox);
             assert_eq!(src_box.box_size(), header.size);
 
@@ -124,7 +125,7 @@ mod tests {
 
         {
             let mut reader = BufReader::new(Cursor::new(&buf));
-            let header = read_box_header(&mut reader).unwrap();
+            let header = BoxHeader::read(&mut reader).unwrap();
             assert_eq!(header.name, BoxType::SttsBox);
             assert_eq!(src_box.box_size(), header.size);
