+++ src/atoms/tkhd.rs
@@ -3,6 +3,7 @@ use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use num_rational::Ratio;
 
 use crate::*;
+use crate::atoms::*;
 
 
 #[derive(Debug, PartialEq)]
@@ -140,7 +141,7 @@ impl<R: Read + Seek> ReadBox<&mut BufReader<R>> for TkhdBox {
 impl<W: Write> WriteBox<&mut BufWriter<W>> for TkhdBox {
     fn write_box(&self, writer: &mut BufWriter<W>) -> Result<u64> {
         let size = self.box_size();
-        BoxHeader::new(Self::box_type(), size).write_box(writer)?;
+        BoxHeader::new(Self::box_type(), size).write(writer)?;
 
         write_box_header_ext(writer, self.version, self.flags)?;
 
@@ -186,7 +187,7 @@ impl<W: Write> WriteBox<&mut BufWriter<W>> for TkhdBox {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::read_box_header;
+    use crate::atoms::BoxHeader;
     use std::io::Cursor;
 
     #[test]
@@ -224,7 +225,7 @@ mod tests {
 
         {
             let mut reader = BufReader::new(Cursor::new(&buf));
-            let header = read_box_header(&mut reader).unwrap();
+            let header = BoxHeader::read(&mut reader).unwrap();
             assert_eq!(header.name, BoxType::TkhdBox);
             assert_eq!(src_box.box_size(), header.size);
 
@@ -269,7 +270,7 @@ mod tests {
 
         {
             let mut reader = BufReader::new(Cursor::new(&buf));
-            let header = read_box_header(&mut reader).unwrap();
+            let header = BoxHeader::read(&mut reader).unwrap();
             assert_eq!(header.name, BoxType::TkhdBox);
             assert_eq!(src_box.box_size(), header.size);
